refactor v4 client
