+++ packages/js/src/fabric/Conversation.ts
@@ -1,7 +1,6 @@
 import { type ConversationEventParams } from '@signalwire/core'
 import { HTTPClient } from './HTTPClient'
 import { WSClient } from './WSClient'
-import { WSClientV4 } from './v4/WSClientV4'
 import type {
   GetConversationsResponse,
   GetMessagesParams,
@@ -35,13 +34,13 @@ const DEFAULT_CHAT_MESSAGES_PAGE_SIZE = 10
 
 interface ConversationOptions {
   httpClient: HTTPClient
-  wsClient: WSClient | WSClientV4
+  wsClient: WSClient
 }
 
 // TODO: Implement a TS contract
 export class Conversation {
   private httpClient: HTTPClient
-  private wsClient: WSClient | WSClientV4
+  private wsClient: WSClient
   private callbacks = new Set<ConversationSubscribeCallback>()
   private chatSubscriptions: Record<
     string,
